# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Food-Delivery",
    "description": "<b>RESPONSES</b>::\n 200-> OK\n   401->Not Authorized\n   201->No Record\n 400->Bad Request \n 500->Internal Server Error",
    
      "version": "1.0"
  },
  "produces": ["application/json"],
  "paths": {
...
   "/api/mobile/company/getCompanies": {
     "get": {
       "x-swagger-router-controller": "/api/mobile/company/getCompanies",
       "operationId": "index",
       "tags": ["All Vendors listing"],
       "description": "",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
         { "name": "deliveryType",  "in": "query",  "required":"true","type":"string"},
         { "name": "page",  "in": "query","type":"string"},
         { "name": "limit",  "in": "query","type":"string"},
         { "name": "lat",  "in": "query","type":"string"},
         { "name": "lng",  "in": "query","type":"string"},
         { "name": "itemType",  "in": "query","type":"string"}



    

    ],

  
   
   
       "responses": {}
     }
   },
   "/api/mobile/company/detail": {
     "get": {
       "x-swagger-router-controller": "/api/mobile/company/detail",
       "operationId": "index",
       "tags": ["Get Vendor Detail"],
       "description": "",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
         { "name": "companyId",  "in": "query",  "required":"true","type":"string"}


    ],

  
   
   
       "responses": {}
     }
   },
   "/api/mobile/company/shipmentCharges": {
     "post": {
       "x-swagger-router-controller": "/api/mobile/company/shipmentCharges",
       "operationId": "index",
       "tags": ["Get Shipment Charges"],
       "description": "",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
         { "name": "companyId",  "in": "formData",  "required":"true","type":"string"},
         { "name": "addressId",  "in": "formData",  "required":"true","type":"string"}


    ],

  
   
   
       "responses": {}
     }
   },
   "/api/mobile/company/gallery": {
    "post": {
      "x-swagger-router-controller": "/api/mobile/company/gallery",
      "operationId": "index",
      "tags": ["Get Gallery"],
      "description": "",
      "parameters": [
        { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
        { "name": "companyId",  "in": "formData",  "required":"true","type":"string"},
        { "name": "mediaType",  "in": "formData","type":"string" ,"description":"send 'photo' to get photos,'video' to get videos and send '' for all type of files "},
        { "name": "page",  "in": "formData", "type":"string"},
        { "name": "limit",  "in": "formData","type":"string"}


   ],

 
  
  
      "responses": {}
    }
  },
  "/api/mobile/company/gallery/add": {
    "post": {
      "x-swagger-router-controller": "/api/mobile/company/gallery/add",
      "operationId": "index",
      "tags": ["Add  Gallery images"],
      "description": "",
      "parameters": [
        { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
        { "name": "companyId",  "in": "formData",  "required":"true","type":"string"},
        { "name": "mediaType",  "required":"true", "in": "formData","type":"string" ,"description":"send 'photo' to upload photos,'video' to uplaod videos"},
        { "name": "title",  "required":"true", "in": "formData", "type":"string"},
        { "name": "description", "required":"true", "in": "formData","type":"string"}


   ],

 
  
  
      "responses": {}
    }
  },
...
   "/api/mobile/coupan/getPromoList": {
     "get": {
       "x-swagger-router-controller": "/api/mobile/coupan/getPromoList",
       "operationId": "index",
       "tags": ["Get Promo Listing"],
       "description": "Promo Listing",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"}


    ],

  
   
   
       "responses": {}
     }
   },
   "/api/mobile/coupan/applyCoupan": {
     "post": {
       "x-swagger-router-controller": "/api/mobile/coupan/applyCoupan",
       "operationId": "index",
       "tags": ["Apply Coupon"],
       "description": "Apply Coupon",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
         { "name": "promoCode",  "in": "formData",  "required":"true" ,"type":"string"}


    ],

       "responses": {}
     }
   },
   "/api/mobile/coupan/removeCoupan": {
     "post": {
       "x-swagger-router-controller": "/api/mobile/coupan/removeCoupan",
       "operationId": "index",
       "tags": ["Remove applied  Coupon"],
       "description": "Remove applied Coupon",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
         { "name": "promoCode",  "in": "formData",  "required":"true" ,"type":"string"}


    ],

       "responses": {}
     }
   },
...
  }
}